+++ c-scape/src/lib.rs
@@ -58,7 +58,6 @@ use error_str::error_str;
 use libc::c_ulong;
 use libc::{c_char, c_int, c_long};
 use rustix::ffi::CStr;
-use rustix::fs::{AtFlags, Mode, OFlags};
 
 // ctype
 mod ctype;
@@ -604,92 +603,6 @@ unsafe extern "C" fn posix_spawn_file_actions_init(_ptr: *const c_void) -> c_int
 }
 
 // exec
-
-#[cfg(any(target_os = "android", target_os = "linux"))]
-#[no_mangle]
-unsafe extern "C" fn execvp(file: *const c_char, args: *const *const c_char) -> c_int {
-    libc!(libc::execvp(file, args));
-
-    let cwd = rustix::fs::cwd();
-
-    let file = CStr::from_ptr(file);
-    if file.to_bytes().contains(&b'/') {
-        let err = rustix::runtime::execve(file, args.cast(), environ.cast());
-        set_errno(Errno(err.raw_os_error()));
-        return -1;
-    }
-
-    let path = _getenv(rustix::cstr!("PATH"));
-    let path = if path.is_null() {
-        rustix::cstr!("/bin:/usr/bin")
-    } else {
-        CStr::from_ptr(path)
-    };
-
-    let mut access_error = false;
-    for dir in path.to_bytes().split(|byte| *byte == b':') {
-        // Open the directory and use `execveat` to avoid needing to
-        // dynamically allocate a combined path string ourselves.
-        let result = rustix::fs::openat(
-            cwd,
-            dir,
-            OFlags::PATH | OFlags::DIRECTORY | OFlags::CLOEXEC | OFlags::NOCTTY,
-            Mode::empty(),
-        )
-        .and_then::<(), _>(|dirfd| {
-            let mut error = rustix::runtime::execveat(
-                &dirfd,
-                file,
-                args.cast(),
-                environ.cast(),
-                AtFlags::empty(),
-            );
-
-            // If `execveat` is unsupported, emulate it with `execve`, without
-            // allocating. This trusts /proc/self/fd.
-            #[cfg(any(target_os = "android", target_os = "linux"))]
-            if let rustix::io::Errno::NOSYS = error {
-                let fd = rustix::fs::openat(
-                    &dirfd,
-                    file,
-                    OFlags::PATH | OFlags::CLOEXEC | OFlags::NOCTTY,
-                    Mode::empty(),
-                )?;
-                const PREFIX: &[u8] = b"/proc/self/fd/";
-                const PREFIX_LEN: usize = PREFIX.len();
-                let mut buf = [0_u8; PREFIX_LEN + 20 + 1];
-                buf[..PREFIX_LEN].copy_from_slice(PREFIX);
-                let fd_dec = rustix::path::DecInt::from_fd(&fd);
-                let fd_bytes = fd_dec.as_c_str().to_bytes_with_nul();
-                buf[PREFIX_LEN..PREFIX_LEN + fd_bytes.len()].copy_from_slice(fd_bytes);
-                error = rustix::runtime::execve(
-                    CStr::from_bytes_with_nul_unchecked(&buf),
-                    args.cast(),
-                    environ.cast(),
-                );
-            }
-
-            Err(error)
-        });
-
-        match result {
-            Ok(()) => (),
-            Err(rustix::io::Errno::ACCESS) => access_error = true,
-            Err(rustix::io::Errno::NOENT) | Err(rustix::io::Errno::NOTDIR) => {}
-            Err(err) => {
-                set_errno(Errno(err.raw_os_error()));
-                return -1;
-            }
-        }
-    }
-
-    set_errno(Errno(if access_error {
-        libc::EACCES
-    } else {
-        libc::ENOENT
-    }));
-    -1
-}
 
 #[cfg(not(target_os = "wasi"))]
 #[no_mangle]
